[minecraft-format-codes] add a method to specify default attributes/styles and ones that appear in all tags
Use the "all" key to specify attributes/styles that appear in all tags,
and "default" to specify a default.

If an "all" value is given and either another attribute/style is given
for a tag or a default is given, then both values will be concatenated
to each other, joined by a space.