[bnaystream] Add support for boolean options in the config file
If an option name has no equals sign and value after it, then it will be
treated as a boolean value.  Option names starting with `no-` or `no_`
will have that prefix removed and their values set to 0; other boolean
option names will be left intact and have their values set to 1.