[mini-browser] Allow setting the User-Agent header
This also clears the cache if necessary to prevent an old user agent
from being used:  a file called "hash" is stored in the cache directory,
containing a SHA-512 hash of the WM_CLASS hint and the user-provided
user agent (or empty string if unset), concatenated by a zero byte.
The hash is computed on startup and compared with the hash file; if
they do not match (or if the file does not exist), the cache is cleared.