[apic-{extract,set}, audiobnay] Relicense to GPLv3+ because of Mutagen dependency
Mutagen has always been licensed under the GPL, but I did not check when
I began using Mutagen in (or prior to, but probably in) 2010 (when I was
at most 18).

I've added an exception to allow use under the X11 License if the
Mutagen dependency is removed, or if a portion which does not depend
on Mutagen is used.

The above arrangement (i.e. the license headers added in this commit)
also applies to the versions of the programs relicensed in this commit
in all previous commits to the extent possible, except that prior to
Mutagen commit 598f0be6b5ed0fe18e8bf0560c54a9c070754d79 (2017-02-13),
Mutagen was licensed under GPLv2 only, and so that license replaces
GPLv3 or later when one of the programs relicensed in this commit uses
versions of Mutagen that were published prior to that commit.

To be clear, the programs relicensed in this commit are apic-extract,
apic-set, and audiobnay.  The licenses of all other programs and other
files in this repository remain unchanged.

My apologies to the Mutagen developers.