+import - stop some Python scripts misisng shebang lines from running as shell scripts
This attempts to determine if the parent process is a Python script
erroneously running in a shell, and if so, kills the parent with
SIGKILL.  Otherwise, execution will pass through to the real import(1).

This will not detect all Python scripts or shells.  It will detect
Python scripts which start with a "from" or "import" statement (ignoring
comments and bare strings), shells with names four characters or
less ending in "sh", and the shell defined by the $SHELL environment
variable.