[wkimgcss] run xvfb-run in a with block instead of execing into it
This uses a context manager which will run a dummy process with xvfb-run
and terminate the dummy process when the context manager is exited (or
when Python exits).  (Terminating xvfb-run itself will result in it not
cleaning up after itself.)

xvfb-run is used instead of Xvfb because the former will configure
authentication for the server and find a free display number, instead
of having to do those manually.